Add server failing handling for apps

+++ server/controllers/workspace.py
@@ -2,6 +2,7 @@ import json
 import os
 import shutil
 import uuid
+import tempfile
 
 from fastapi import HTTPException
 
@@ -25,13 +26,17 @@ class WorkspaceFolderController:
         self.r_path_to_workspace = r_path_to_workspace
 
     def write_workspace_properties(self, workspace_properties: dict):
-        workspace_properties_path = os.path.join(self.r_path_to_workspace, "properties.json")
+        workspace_properties_path = os.path.join(
+            self.r_path_to_workspace, "properties.json"
+        )
         with open(workspace_properties_path, "w") as file:
             json.dump(workspace_properties, file, indent=2)
 
     def get_workspace_properties(self):
         if os.path.exists(os.path.join(self.r_path_to_workspace, "properties.json")):
-            with open(os.path.join(self.r_path_to_workspace, "properties.json"), "r") as file:
+            with open(
+                os.path.join(self.r_path_to_workspace, "properties.json"), "r"
+            ) as file:
 
                 props = json.load(file)
                 return props.get("apps", [])
@@ -58,13 +63,16 @@ class WorkspaceFolderController:
         target_app = self.get_app(app_id=app_id)
         if target_app is None:
             raise HTTPException(
-                status_code=400, detail="App does not exist, or id does not exist for this app."
+                status_code=400,
+                detail="App does not exist, or id does not exist for this app.",
             )
 
         existing_app_labels = [a["label"] for a in self.get_workspace_properties()]
 
         if new_label in existing_app_labels:
-            raise HTTPException(status_code=400, detail="Another app with the same label already exists")
+            raise HTTPException(
+                status_code=400, detail="Another app with the same label already exists"
+            )
 
         app_info = {**target_app, "label": new_label}
         workspace_data = self.get_workspace_properties()
@@ -101,14 +109,18 @@ class AppFolderController:
             json.dump(app_properties_data, file, indent=2)
 
     def _get_workspace_properties(self):
-        workspace_properties_path = os.path.join(self.r_path_to_workspace, "properties.json")
+        workspace_properties_path = os.path.join(
+            self.r_path_to_workspace, "properties.json"
+        )
         if not os.path.exists(workspace_properties_path):
             return None
         with open(workspace_properties_path, "r") as file:
             return json.load(file)
 
     def _write_workspace_properties(self, workspace_properties: dict):
-        workspace_properties_path = os.path.join(self.r_path_to_workspace, "properties.json")
+        workspace_properties_path = os.path.join(
+            self.r_path_to_workspace, "properties.json"
+        )
         with open(workspace_properties_path, "w") as file:
             json.dump(workspace_properties, file, indent=2)
 
@@ -143,6 +155,14 @@ class AppFolderController:
         self._write_workspace_properties(workspace_properties)
         return app_object
 
+    def _remove_app_from_workspace_properties(self, app_name: str):
+        workspace_properties = self._get_workspace_properties()
+        for app in workspace_properties["apps"]:
+            if app["name"] == app_name:
+                workspace_properties["apps"].remove(app)
+                break
+        self._write_workspace_properties(workspace_properties)
+
     def _create_default_workspace_files(
         self, router: DropbaseRouter, app_label: str = None
     ) -> str | None:
@@ -162,13 +182,19 @@ class AppFolderController:
                 file_name="properties.json",
             )
             if router:
-                router.app.create_app(app_properties={**app_object})
+                response = router.app.create_app(app_properties={**app_object})
+
+            if response.status_code != 200:
+                shutil.rmtree(self.app_folder_path)
+                self._remove_app_from_workspace_properties(self.app_name)
+                raise HTTPException(
+                    status_code=500, detail="Unable to create app folder"
+                )
 
             # Create new page folder with __init__.py
             self.create_page(router=router)
 
-        except Exception as e:
-            print("Unable to create app folder", e)
+        except Exception:
             raise HTTPException(status_code=500, detail="Unable to create app folder")
 
     def _get_app_properties(self):
@@ -191,7 +217,9 @@ class AppFolderController:
         created_app_names = get_subdirectories(self.r_path_to_workspace)
         app_info = []
         for app_name in created_app_names:
-            app_properties = os.path.join(self.r_path_to_workspace, app_name, "properties.json")
+            app_properties = os.path.join(
+                self.r_path_to_workspace, app_name, "properties.json"
+            )
             if not os.path.exists(app_properties):
                 app_info.append({"name": app_name, "label": app_name, "id": None})
                 continue
@@ -304,7 +332,8 @@ class AppFolderController:
 
         if len(app_properties_data["pages"]) == 1:
             raise HTTPException(
-                status_code=400, detail="Cannot delete the only page in the app. Delete the app instead."
+                status_code=400,
+                detail="Cannot delete the only page in the app. Delete the app instead.",
             )
 
         # delete page folder
@@ -340,13 +369,19 @@ class AppFolderController:
             )
 
         if check_if_object_exists(self.app_folder_path):
-            raise HTTPException(status_code=400, detail="Another app with the same name already exists")
+            raise HTTPException(
+                status_code=400, detail="Another app with the same name already exists"
+            )
 
         self.create_workspace_properties()
 
-        existing_app_labels = [a["label"] for a in self._get_workspace_properties()["apps"]]
+        existing_app_labels = [
+            a["label"] for a in self._get_workspace_properties()["apps"]
+        ]
         if app_label in existing_app_labels:
-            raise HTTPException(status_code=400, detail="Another app with the same label already exists")
+            raise HTTPException(
+                status_code=400, detail="Another app with the same label already exists"
+            )
 
         self._create_default_workspace_files(router=router, app_label=app_label)
 
@@ -354,25 +389,46 @@ class AppFolderController:
 
     def delete_app(self, app_name: str, router: DropbaseRouter):
         app_path = os.path.join(self.r_path_to_workspace, app_name)
-        if os.path.exists(app_path):
-            shutil.rmtree(app_path)
 
-            workspace_properties = self._get_workspace_properties()
-            if not workspace_properties:
+        with tempfile.NamedTemporaryFile() as temp_workspace_file:
+            workspace_properties_path = os.path.join(
+                self.r_path_to_workspace, "properties.json"
+            )
+            if not os.path.exists(workspace_properties_path):
                 return
 
-            app_id = None
-            for app in workspace_properties["apps"]:
-                if app["name"] == app_name:
-                    app_id = app.get("id", None)
-                    workspace_properties["apps"].remove(app)
-                    break
+            shutil.copy2(workspace_properties_path, temp_workspace_file.name)
 
-            self._write_workspace_properties(workspace_properties)
-            router.app.delete_app(app_id=app_id)
-            return {"message": "App deleted"}
-        else:
-            raise HTTPException(status_code=400, detail="App does not exist")
+            with tempfile.TemporaryDirectory() as backup_dir:
+                if os.path.exists(app_path):
+                    shutil.copytree(app_path, backup_dir, dirs_exist_ok=True)
+                    workspace_properties = self._get_workspace_properties()
+                    if not workspace_properties:
+                        return
+
+                    app_id = None
+                    for app in workspace_properties["apps"]:
+                        if app["name"] == app_name:
+                            app_id = app.get("id", None)
+                            workspace_properties["apps"].remove(app)
+                            break
+
+                    self._write_workspace_properties(workspace_properties)
+                    delete_response = router.app.delete_app(app_id=app_id)
+                    if delete_response.status_code != 200:
+                        shutil.rmtree(app_path)
+                        shutil.copytree(backup_dir, app_path)
+                        shutil.copy(temp_workspace_file.name, workspace_properties_path)
+
+                        raise HTTPException(
+                            status_code=500, detail="Unable to delete app folder"
+                        )
+
+                    shutil.rmtree(app_path)
+
+                    return {"message": "App deleted"}
+                else:
+                    raise HTTPException(status_code=400, detail="App does not exist")
 
 
 INIT_CODE = """
